made everything into using packages and (somewhat) of a logical structure - still a lot of cleanup necessary for relations/interconnections to be logical/obvious. sfx -> resource folder, gfx and maps -> normal folders. Main class: game2d.core.GameWindow
