+++ src/game2d/character/charmechanics/charstats/Equipment.java
@@ -1,0 +1,18 @@
+package game2d.character.charmechanics.charstats;
+
+public class Equipment{
+	/**
+	 * 0: head
+	 * <br>1: arms
+	 * <br>2: chest
+	 * <br>3: legs
+	 * <br>4: neck
+	 * <br>5: ring*/
+	private int slot;
+	private String name;
+
+	public Equipment(String name, int slot){
+		this.name = name;
+		this.slot = slot;
+	}
+}
