+++ src/game2d/util/shapes/Point.java
@@ -1,0 +1,15 @@
+package game2d.util.shapes;
+
+public class Point{
+	public int x, y;
+
+	public Point(int x, int y){
+		this.x = x;
+		this.y = y;
+	}
+
+	public Point(){
+		x = 0;
+		y = 0;
+	}
+}
